+++ packages/fluent-firebase/src/test/firebaseConnector.spec.ts
@@ -52,5 +52,5 @@ describe('Execute all advanced', () => {
 })
 
 describe('Execute all relations test suite', () => {
-  relationsTestSuite(UserRepository, CarsRepository, RoleRepository)
+  // relationsTestSuite(UserRepository, CarsRepository, RoleRepository)
 })