+++ packages/fluent-firebase/src/test/user.repository.ts
@@ -1,8 +1,9 @@
 import { CarsEntity } from '@goatlab/fluent/src/TypeOrmConnector/test/relations/car/car.entity'
 import { UsersEntity } from '@goatlab/fluent/src/TypeOrmConnector/test/relations/user/user.entity'
 import {
-  UsersEntityInputSchema,
-  UsersEntitySchema
+  userInputSchema,
+  UsersDtoIn,
+  UsersDtoOut
 } from '@goatlab/fluent/src/TypeOrmConnector/test/relations/user/user.schema'
 import { CarsRepository } from './car.repository'
 import { FirebaseConnector } from '../FirebaseConnector'
@@ -11,12 +12,13 @@ import { RoleUsersRepository } from './roles_user.repository'
 
 export class UserRepository extends FirebaseConnector<
   UsersEntity,
-  UsersEntityInputSchema
+  UsersDtoIn,
+  UsersDtoOut
 > {
   constructor() {
     super({
       entity: UsersEntity,
-      inputSchema: UsersEntitySchema
+      inputSchema: userInputSchema
     })
   }