+++ packages/fluent/src/TypeOrmConnector/queryBuilder/mongodb/getMongoWhere.ts
@@ -12,8 +12,6 @@ import {
 import { extractConditions } from '../../util/extractConditions'
 
 export interface getTypeOrmMongoWhereParams {
-
-
   where?: FluentQuery<AnyObject>['where']
 }
 
@@ -29,8 +27,8 @@ export const getMongoWhere = ({
     return {}
   }
 
-  const Filters: { where: { $or: any[] } } = {
-    where: { $or: [{ $and: [] }] }
+  const Filters: { filter: { $or: any[] } } = {
+    filter: { $or: [{ $and: [] }] }
   }
 
   const orConditions = extractConditions(where['OR'])
@@ -62,39 +60,39 @@ export const getMongoWhere = ({
 
     switch (operator) {
       case LogicOperator.equals:
-        Filters.where.$or[0].$and.push({ [element]: { $eq: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $eq: value } })
         break
       case LogicOperator.isNot:
-        Filters.where.$or[0].$and.push({ [element]: { $neq: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $neq: value } })
         break
       case LogicOperator.greaterThan:
-        Filters.where.$or[0].$and.push({ [element]: { $gt: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $gt: value } })
         break
       case LogicOperator.greaterOrEqualThan:
-        Filters.where.$or[0].$and.push({ [element]: { $gte: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $gte: value } })
         break
       case LogicOperator.lessThan:
-        Filters.where.$or[0].$and.push({ [element]: { $lt: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $lt: value } })
         break
       case LogicOperator.lessOrEqualThan:
-        Filters.where.$or[0].$and.push({ [element]: { $lte: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $lte: value } })
         break
       case LogicOperator.in:
-        Filters.where.$or[0].$and.push({ [element]: { $in: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $in: value } })
         break
       case LogicOperator.notIn:
-        Filters.where.$or[0].$and.push({
+        Filters.filter.$or[0].$and.push({
           [element]: { $not: { $in: value } }
         })
         break
       case LogicOperator.exists:
-        Filters.where.$or[0].$and.push({ [element]: { $exists: true } })
+        Filters.filter.$or[0].$and.push({ [element]: { $exists: true } })
         break
       case LogicOperator.notExists:
-        Filters.where.$or[0].$and.push({ [element]: { $exists: false } })
+        Filters.filter.$or[0].$and.push({ [element]: { $exists: false } })
         break
       case LogicOperator.regexp:
-        Filters.where.$or[0].$and.push({ [element]: { $regex: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $regex: value } })
         break
     }
   }
@@ -114,39 +112,39 @@ export const getMongoWhere = ({
 
     switch (operator) {
       case LogicOperator.equals:
-        Filters.where.$or[0].$and.push({ [element]: { $eq: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $eq: value } })
         break
       case LogicOperator.isNot:
-        Filters.where.$or[0].$and.push({ [element]: { $neq: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $neq: value } })
         break
       case LogicOperator.greaterThan:
-        Filters.where.$or[0].$and.push({ [element]: { $gt: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $gt: value } })
         break
       case LogicOperator.greaterOrEqualThan:
-        Filters.where.$or[0].$and.push({ [element]: { $gte: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $gte: value } })
         break
       case LogicOperator.lessThan:
-        Filters.where.$or[0].$and.push({ [element]: { $lt: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $lt: value } })
         break
       case LogicOperator.lessOrEqualThan:
-        Filters.where.$or[0].$and.push({ [element]: { $lte: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $lte: value } })
         break
       case LogicOperator.in:
-        Filters.where.$or[0].$and.push({ [element]: { $in: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $in: value } })
         break
       case LogicOperator.notIn:
-        Filters.where.$or[0].$and.push({
+        Filters.filter.$or[0].$and.push({
           [element]: { $not: { $in: value } }
         })
         break
       case LogicOperator.exists:
-        Filters.where.$or[0].$and.push({ [element]: { $exists: true } })
+        Filters.filter.$or[0].$and.push({ [element]: { $exists: true } })
         break
       case LogicOperator.notExists:
-        Filters.where.$or[0].$and.push({ [element]: { $exists: false } })
+        Filters.filter.$or[0].$and.push({ [element]: { $exists: false } })
         break
       case LogicOperator.regexp:
-        Filters.where.$or[0].$and.push({ [element]: { $regex: value } })
+        Filters.filter.$or[0].$and.push({ [element]: { $regex: value } })
         break
     }
   }
@@ -166,44 +164,44 @@ export const getMongoWhere = ({
 
     switch (operator) {
       case LogicOperator.equals:
-        Filters.where.$or.push({ [element]: { $eq: value } })
+        Filters.filter.$or.push({ [element]: { $eq: value } })
         break
       case LogicOperator.isNot:
-        Filters.where.$or.push({ [element]: { $neq: value } })
+        Filters.filter.$or.push({ [element]: { $neq: value } })
         break
       case LogicOperator.greaterThan:
-        Filters.where.$or.push({ [element]: { $gt: value } })
+        Filters.filter.$or.push({ [element]: { $gt: value } })
         break
       case LogicOperator.greaterOrEqualThan:
-        Filters.where.$or.push({ [element]: { $gte: value } })
+        Filters.filter.$or.push({ [element]: { $gte: value } })
         break
       case LogicOperator.lessThan:
-        Filters.where.$or.push({ [element]: { $lt: value } })
+        Filters.filter.$or.push({ [element]: { $lt: value } })
         break
       case LogicOperator.lessOrEqualThan:
-        Filters.where.$or.push({ [element]: { $lte: value } })
+        Filters.filter.$or.push({ [element]: { $lte: value } })
         break
       case LogicOperator.in:
-        Filters.where.$or.push({ [element]: { $in: value } })
+        Filters.filter.$or.push({ [element]: { $in: value } })
         break
       case LogicOperator.notIn:
-        Filters.where.$or.push({
+        Filters.filter.$or.push({
           [element]: { $not: { $in: value } }
         })
         break
       case LogicOperator.exists:
-        Filters.where.$or.push({ [element]: { $exists: true } })
+        Filters.filter.$or.push({ [element]: { $exists: true } })
         break
       case LogicOperator.notExists:
-        Filters.where.$or.push({ [element]: { $exists: false } })
+        Filters.filter.$or.push({ [element]: { $exists: false } })
         break
       case LogicOperator.regexp:
-        Filters.where.$or.push({ [element]: { $regex: value } })
+        Filters.filter.$or.push({ [element]: { $regex: value } })
         break
     }
   }
 
-  const filtered = clearEmpties(Filters.where)
+  const filtered = clearEmpties(Filters.filter)
 
   return filtered
 }